+++ SKLibrary/src/main/java/com/silvericekey/skutilslibrary/utils/HttpUtil.kt
@@ -1,8 +1,12 @@
 package com.silvericekey.skutilslibrary.utils
 
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkInfo
 import android.text.TextUtils
 import android.util.Log
 import com.silvericekey.skutilslibrary.SKUtilsLibrary
+import com.silvericekey.skutilslibrary.interpolator.CustomCacheInterceptor
 import com.silvericekey.skutilslibrary.net.AddCookiesInterceptor
 import com.silvericekey.skutilslibrary.net.ReceivedCookiesInterceptor
 import okhttp3.*
@@ -16,6 +20,10 @@ import java.io.File
 import java.util.*
 import java.util.concurrent.TimeUnit
 
+/**
+ * 网络请求工具类
+ * 单例模式
+ * */
 class HttpUtil {
     private var okHttpClient: OkHttpClient
     private var okHttpWebSocketClient: OkHttpClient
@@ -31,6 +39,21 @@ class HttpUtil {
             return httpUtils!!
         }
 
+        /**
+         * 是否有网络检测
+         * */
+        fun hasNetwork(): Boolean? {
+            var isConnected: Boolean? = false // Initial Value
+            val connectivityManager = SKUtilsLibrary.context!!.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+            val activeNetwork: NetworkInfo? = connectivityManager.activeNetworkInfo
+            if (activeNetwork != null && activeNetwork.isConnected)
+                isConnected = true
+            return isConnected
+        }
+
+        /**
+         * 初始化连接host
+         * */
         @JvmStatic
         fun init(baseUrl: String) {
             if (httpUtils == null) {
@@ -42,6 +65,9 @@ class HttpUtil {
             }
         }
 
+        /**
+         * websocket测试
+         * */
         @JvmStatic
         fun webSocketTest() {
             var wsUrl = "wss://echo.websocket.org/"
@@ -116,7 +142,9 @@ class HttpUtil {
             }).request()
         }
     }
-
+    /**
+     * 初始化网络
+     * */
     constructor(baseUrl: String) {
         if (TextUtils.isEmpty(baseUrl)) {
             throw Exception("Please set base url first")
@@ -129,8 +157,8 @@ class HttpUtil {
         val receivedCookiesInterceptor = ReceivedCookiesInterceptor()
         val addCookiesInterceptor = AddCookiesInterceptor()
         //缓存文件夹
-        val cacheFile = File(SKUtilsLibrary.context!!.getExternalCacheDir().toString(), "cache")
-        if (!cacheFile.exists()){
+        val cacheFile = File(SKUtilsLibrary.context!!.externalCacheDir, "")
+        if (!cacheFile.exists()) {
             cacheFile.mkdirs()
         }
         //缓存大小为10M
@@ -151,6 +179,7 @@ class HttpUtil {
                 .addInterceptor(loggingInterceptor)
                 .addInterceptor(addCookiesInterceptor)
                 .addInterceptor(receivedCookiesInterceptor)
+                .addInterceptor(CustomCacheInterceptor())
                 .cache(cache)
                 .connectTimeout(10000L, TimeUnit.MILLISECONDS)
                 .readTimeout(10000L, TimeUnit.MILLISECONDS)
@@ -165,22 +194,34 @@ class HttpUtil {
                 .build()
     }
 
+    /**
+     * 添加拦截器
+     * */
     fun addInterceptor(interceptor: Interceptor): HttpUtil {
         okHttpClient = okHttpClient.newBuilder().addInterceptor(interceptor).build()
         retrofit = retrofit.newBuilder().client(okHttpClient).build()
         return this
     }
 
+    /**
+     * 修改基础连接host
+     * */
     fun changeUrl(baseUrl: String): HttpUtil {
         retrofit = retrofit.newBuilder().baseUrl(baseUrl).client(okHttpClient).build()
         return this
     }
 
+    /**
+     * 创建websocket
+     * */
     fun webSocket(url: String, listener: WebSocketListener) {
         var request = Request.Builder().url(url).build()
         okHttpWebSocketClient.newWebSocket(request, listener)
     }
 
+    /**
+     * api类的转换
+     * */
     fun <T> obtainClass(clz: Class<T>): T {
         return retrofit.create(clz)
     }