+++ SKLibrary/src/main/java/com/silvericekey/skutilslibrary/utils/PictureSelectorUtil.kt
@@ -1,0 +1,59 @@
+package com.silvericekey.skutilslibrary.utils
+
+import android.Manifest
+import android.content.Context
+import android.content.Intent
+import android.provider.MediaStore
+import com.blankj.utilcode.util.PermissionUtils
+import com.blankj.utilcode.util.ToastUtils
+import com.silvericekey.skutilslibrary.SKUtilsLibrary
+
+object PictureSelectorUtil {
+    /**
+     * 直接调用系统相机拍照
+     * **/
+    fun getPictureWithCamera(context:Context) {
+        if (getPermission({
+                    ToastUtils.showShort("您拒绝了开启权限，改功能无法使用")
+                }, {
+                    getPictureWithCamera(context)
+                })) {
+            var intent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
+            context.startActivity(intent)
+        }
+    }
+    /*
+    * 调用系统相册
+    * */
+    fun getPictureWithGallery(context: Context) {
+        if (getPermission({
+                    ToastUtils.showShort("您拒绝了开启权限，改功能无法使用")
+                }, {
+                    getPictureWithGallery(context)
+                })) {
+            var intent = Intent(Intent.ACTION_GET_CONTENT)
+            intent.setType("image/*");
+
+            context.startActivity(intent)
+        }
+    }
+
+    private fun getPermission(onGranted: () -> Unit, onDenied: () -> Unit): Boolean {
+        if (!PermissionUtils.isGranted(Manifest.permission.WRITE_EXTERNAL_STORAGE, Manifest.permission.CAMERA)) {
+            PermissionUtils.permission(Manifest.permission.WRITE_EXTERNAL_STORAGE, Manifest.permission.CAMERA)
+                    .callback(object : PermissionUtils.SimpleCallback {
+                        override fun onGranted() {
+                            onGranted()
+                        }
+
+                        override fun onDenied() {
+                            onDenied()
+                        }
+
+                    }).request()
+            return false
+        } else {
+            return true
+        }
+    }
+}