+++ app/src/main/java/com/silverknife/meizhi/mvp/presenter/FunctionAnimMovePresenter.kt
@@ -1,0 +1,10 @@
+package com.silverknife.meizhi.mvp.presenter
+
+import com.silvericekey.skutilslibrary.base.BasePresenter
+
+class FunctionAnimMovePresenter: BasePresenter() {
+    override fun onDestroy() {
+
+    }
+
+}