+++ app/src/main/java/com/silverknife/meizhi/mvp/presenter/RNPresenter.kt
@@ -1,0 +1,9 @@
+package com.silverknife.meizhi.mvp.presenter
+
+import com.silvericekey.skutilslibrary.base.BasePresenter
+
+class RNPresenter : BasePresenter() {
+    override fun onDestroy() {
+
+    }
+}